notes_api/requirements.txt
# Create a requirements.txt file in your Django project's root directory, including the required packages:
Django
djangorestframework
django-cors-headers